Remove unnecessary collecting of checkItems
I think this was necesssary until 5c0f4f6821f1a702a5d5f53ee9fb0320c51044cd, but since then has not been required. Before that commit the `items` were collected for _all_ checklists and then the `Promise.all(items)` ensured all `item` promises were resolved before continuing. In that commit the `reduce` was introduced (to reduce the rate at which we're hitting the Trello API - see the PR) which meant that a _chain_ of promises was generated, one for each checklist (the items for a single checklist still being covered by a `Promise.all()`). The fact it's a chain ensures everything is resolved before continuing.